Extract function to own file
